Inventory Copy done

+++ database/migrations/2014_10_12_000000_create_master_kitchen_inventory_table.php
@@ -14,10 +14,10 @@ return new class extends Migration {
             $table->id();
             $table->string('category'); // Category name or ID
             $table->string('item_name'); // Item name
-            $table->integer('quantity')->default(0); // Quantity
+            $table->float('quantity')->default(0); // Quantity
             $table->string('location_id');
             $table->string('unit'); // Unit (e.g., kg, pcs, liters)
-            $table->integer('price'); // Price with 2 decimal places
+            $table->float('price'); // Price with 2 decimal places
             $table->integer('is_active')->default(true);
             $table->integer('is_deleted')->default(false);
             $table->timestamps(); // Created at & Updated at
